requirements.txt
# H2 Seep Detection - Dependencies
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0  # Data augmentation

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Satellite Data
sentinelhub>=3.9.0
earthengine-api>=0.1.370

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web Scraping (for Google Maps)
selenium>=4.10.0
webdriver-manager>=4.0.0
requests>=2.31.0

# API (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Logging & Monitoring
tensorboard>=2.14.0
tqdm>=4.65.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.5.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
